+++ packages/react/src/transactions/useUnfollow.ts
@@ -12,13 +12,24 @@ import { useState } from 'react';
 import { TransactionState, useHasPendingTransaction } from './adapters/TransactionQueuePresenter';
 import { useUnfollowController } from './adapters/useUnfollowController';
 
+export class PrematureUnfollowError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
+
 export type UseUnfollowArgs = {
   profile: ProfileFieldsFragment;
 };
 
 export function useUnfollow({ profile }: UseUnfollowArgs) {
   const [error, setError] = useState<
-    PendingSigningRequestError | WalletConnectionError | UserRejectedError | IEquatableError | null
+    | PendingSigningRequestError
+    | WalletConnectionError
+    | UserRejectedError
+    | IEquatableError
+    | PrematureUnfollowError
+    | null
   >(null);
   const unfollow = useUnfollowController();
 
@@ -30,6 +41,13 @@ export function useUnfollow({ profile }: UseUnfollowArgs) {
 
   return {
     unfollow: async () => {
+      if (!profile.isFollowedByMe && profile.isOptimisticFollowedByMe) {
+        setError(
+          new PrematureUnfollowError(`Your follow request for ${profile.handle} is still pending.`),
+        );
+        return;
+      }
+
       const result = await unfollow({
         kind: TransactionKind.UNFOLLOW_PROFILE,
         profileId: profile.id,